rename SEO component
